requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
colorama==0.4.6
comm==0.2.3
cryptography==46.0.6
debugpy==1.8.20
decorator==5.2.1
dnspython==2.8.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mcp==1.27.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.4.4
ollama==0.6.1
packaging==26.0
pandas==3.0.2
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311
pyzmq==27.1.0
referencing==0.37.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sse-starlette==3.3.4
stack-data==0.6.3
starlette==1.0.0
tornado==6.5.5
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.43.0
wcwidth==0.6.0
